+++ R/testData.R
@@ -1,0 +1,3 @@
+#' Test dataset for the paper
+
+"testData"